Version without saga
